+++ R/ewma.R
@@ -274,13 +274,16 @@ plot.ewma.qcc <- function(x,
     } 
      
   if(missing(title))
-    { if(is.null(newstats))
-         main.title <- paste("EWMA chart for", data.name)
-      else if(chart.all)
-                main.title <- paste("EWMA chart for", data.name, 
-                                  "and", newdata.name)
-           else main.title <- paste("EWMA chart for", newdata.name) }
-  else main.title <- paste(title)
+  { 
+    if(is.null(newstats))
+      title <- paste("EWMA chart for", data.name)
+    else if(chart.all)
+           title <- paste("EWMA chart for", data.name, "and", newdata.name)
+         else 
+           title <- paste("EWMA chart for", newdata.name) 
+  }
+  if(isFALSE(title) | is.na(title)) title <- ""
+
   cex.labels <- par("cex")*qcc.options("cex")
   cex.stats <- par("cex")*qcc.options("cex.stats")
   
@@ -306,9 +309,9 @@ plot.ewma.qcc <- function(x,
        cex.axis = par("cex.axis")*0.9)
   axis(2, las = axes.las, cex.axis = par("cex.axis")*0.9)
   box()
-  top.line <- par("mar")[3] - length(capture.output(cat(main.title)))
+  top.line <- par("mar")[3] - length(capture.output(cat(title)))
   top.line <- top.line - if(chart.all & (!is.null(newstats))) 0.1 else 0.5
-  mtext(main.title, side = 3, outer = TRUE, 
+  mtext(title, side = 3, outer = TRUE, 
         line = 0, adj = 0, at = par("plt")[1],
         font = par("font.main"), 
         cex  = par("cex")*qcc.options("cex"),