+++ R/mqcc.R
@@ -249,13 +249,15 @@ plot.mqcc <- function(x,
      }
   #
   if(missing(title))
-    { if(is.null(newstats))
-         main.title <- paste(type, "chart for", data.name)
+  { 
+    if(is.null(newstats))
+         title <- paste(type, "chart for", data.name)
        else if(chart.all)
-               main.title <- paste(type, "chart for", data.name, 
-                                   "and", newdata.name)
-            else main.title <- paste(type, "chart for", newdata.name) }
-  else main.title <- paste(title)
+               title <- paste(type, "chart for", data.name, "and", newdata.name)
+            else title <- paste(type, "chart for", newdata.name) 
+  }
+  if(isFALSE(title) | is.na(title)) title <- ""
+  
   cex.labels <- par("cex")*qcc.options("cex")
   cex.stats <- par("cex")*qcc.options("cex.stats")
   
@@ -282,7 +284,7 @@ plot.mqcc <- function(x,
        cex.axis = par("cex.axis")*0.9)
   axis(2, las = axes.las, cex.axis = par("cex.axis")*0.9)
   box()
-  mtext(main.title, side = 3, outer = TRUE, 
+  mtext(title, side = 3, outer = TRUE, 
         line = 0, adj = 0, at = par("plt")[1],
         font = par("font.main"), 
         cex  = par("cex")*qcc.options("cex"), 
@@ -456,15 +458,14 @@ ellipseChart <- function(object, chart.all = TRUE, show.id = FALSE, ngrid = 50,
   if(missing(ylim))
      ylim <- range(pretty(stats[,2],1), q2$limits)
   if(missing(title))
-    { if(is.null(object$newstats))
-           main.title <- paste("Ellipse chart for", object$data.name)
-      else if(chart.all)
-                main.title <- paste("Ellipse chart for", 
-                                    object$data.name, "and",
-                                    object$newdata.name)
-           else main.title <- paste("Ellipse chart\nfor", 
-                                    object$newdata.name) }
-  else main.title <- paste(title)
+  { 
+    if(is.null(object$newstats))
+      title <- paste("Ellipse chart for", object$data.name)
+    else if(chart.all)
+      title <- paste("Ellipse chart for", object$data.name, "and", object$newdata.name)
+    else title <- paste("Ellipse chart for", object$newdata.name) 
+  }
+  if(isFALSE(title) | is.na(title)) title <- ""
   #
   grid <- cbind(seq(xlim[1], xlim[2], length = ngrid),
                 seq(ylim[1], ylim[2], length = ngrid))
@@ -492,7 +493,7 @@ ellipseChart <- function(object, chart.all = TRUE, show.id = FALSE, ngrid = 50,
   rect(par("usr")[1], par("usr")[3], par("usr")[2], par("usr")[4], 
        col = qcc.options("bg.figure"))
   box()
-  mtext(main.title, side = 3, outer = TRUE, 
+  mtext(title, side = 3, outer = TRUE, 
         line = 0, adj = 0, at = par("plt")[1],
         font = par("font.main"), 
         cex  = par("cex")*qcc.options("cex"),